# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "UFV API",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/applicators/{applicator}": {
            "get": {
                "tags": [
                    "applicators"
                ],
                "summary": "Display the specified resource.",
                "description": "show applicator",
                "operationId": "App\\Http\\Controllers\\ApplicatorController::show",
                "parameters": [
                    {
                        "name": "applicator",
                        "in": "path",
                        "description": "Applicator id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "only",
                        "in": "query",
                        "description": "filter results using field1;field2;field3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "with",
                        "in": "query",
                        "description": "get relations using relation1;relation2;relation3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show applicator"
                    }
                }
            },
            "put": {
                "tags": [
                    "applicators"
                ],
                "summary": "Update the specified resource in storage",
                "description": "update a applicator on database",
                "operationId": "App\\Http\\Controllers\\ApplicatorController::update",
                "parameters": [
                    {
                        "name": "applicator",
                        "in": "path",
                        "description": "Applicator id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "acronym": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Applicator updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "applicators"
                ],
                "summary": "Remove the specified resource from storage.",
                "description": "remove a applicator on database",
                "operationId": "App\\Http\\Controllers\\ApplicatorController::destroy",
                "parameters": [
                    {
                        "name": "applicator",
                        "in": "path",
                        "description": "Applicator id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Applicator deleted"
                    }
                }
            }
        },
...
        "/categories/{category}": {
            "get": {
                "tags": [
                    "categories"
                ],
                "summary": "Display the specified resource.",
                "description": "show category",
                "operationId": "App\\Http\\Controllers\\CategoryController::show",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "Category id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "only",
                        "in": "query",
                        "description": "filter results using field1;field2;field3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "with",
                        "in": "query",
                        "description": "get relations using relation1;relation2;relation3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show category"
                    }
                }
            },
            "put": {
                "tags": [
                    "categories"
                ],
                "summary": "Update the specified resource in storage",
                "description": "update a category on database",
                "operationId": "App\\Http\\Controllers\\CategoryController::update",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "Category id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Category updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "categories"
                ],
                "summary": "Remove the specified resource from storage.",
                "description": "remove a category on database",
                "operationId": "App\\Http\\Controllers\\CategoryController::destroy",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "Category id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category deleted"
                    }
                }
            }
        },
...
        "/contests/{contest}": {
            "get": {
                "tags": [
                    "contests"
                ],
                "summary": "Display the specified resource.",
                "description": "show contest",
                "operationId": "App\\Http\\Controllers\\ContestController::show",
                "parameters": [
                    {
                        "name": "contest",
                        "in": "path",
                        "description": "Contest id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "only",
                        "in": "query",
                        "description": "filter results using field1;field2;field3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "with",
                        "in": "query",
                        "description": "get relations using relation1;relation2;relation3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show contest"
                    }
                }
            },
            "put": {
                "tags": [
                    "contests"
                ],
                "summary": "Update the specified resource in storage",
                "description": "update a contest on database",
                "operationId": "App\\Http\\Controllers\\ContestController::update",
                "parameters": [
                    {
                        "name": "contest",
                        "in": "path",
                        "description": "Contest id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "year": {
                                        "type": "string"
                                    },
                                    "number": {
                                        "type": "integer"
                                    },
                                    "institution_id": {
                                        "type": "integer"
                                    },
                                    "applicator_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contest updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "contests"
                ],
                "summary": "Remove the specified resource from storage.",
                "description": "remove a contest on database",
                "operationId": "App\\Http\\Controllers\\ContestController::destroy",
                "parameters": [
                    {
                        "name": "contest",
                        "in": "path",
                        "description": "Contest id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contest deleted"
                    }
                }
            }
        },
...
        "/institutions/{institution}": {
            "get": {
                "tags": [
                    "institutions"
                ],
                "summary": "Display the specified resource.",
                "description": "show institution",
                "operationId": "App\\Http\\Controllers\\InstitutionController::show",
                "parameters": [
                    {
                        "name": "institution",
                        "in": "path",
                        "description": "Institution id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "only",
                        "in": "query",
                        "description": "filter results using field1;field2;field3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "with",
                        "in": "query",
                        "description": "get relations using relation1;relation2;relation3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show institution"
                    }
                }
            },
            "put": {
                "tags": [
                    "institutions"
                ],
                "summary": "Update the specified resource in storage",
                "description": "update a institution on database",
                "operationId": "App\\Http\\Controllers\\InstitutionController::update",
                "parameters": [
                    {
                        "name": "institution",
                        "in": "path",
                        "description": "Institution id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "acronym": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Institution updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "institutions"
                ],
                "summary": "Remove the specified resource from storage.",
                "description": "remove a institution on database",
                "operationId": "App\\Http\\Controllers\\InstitutionController::destroy",
                "parameters": [
                    {
                        "name": "institution",
                        "in": "path",
                        "description": "Institution id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Institution deleted"
                    }
                }
            }
        },
...
        "/questions/{question}": {
            "get": {
                "tags": [
                    "questions"
                ],
                "summary": "Display the specified resource.",
                "description": "show question",
                "operationId": "App\\Http\\Controllers\\QuestionController::show",
                "parameters": [
                    {
                        "name": "question",
                        "in": "path",
                        "description": "Question id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "only",
                        "in": "query",
                        "description": "filter results using field1;field2;field3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "with",
                        "in": "query",
                        "description": "get relations using relation1;relation2;relation3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show question"
                    }
                }
            },
            "put": {
                "tags": [
                    "questions"
                ],
                "summary": "Update the specified resource in storage",
                "description": "update a question on database",
                "operationId": "App\\Http\\Controllers\\QuestionController::update",
                "parameters": [
                    {
                        "name": "question",
                        "in": "path",
                        "description": "Question id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "statement": {
                                        "type": "string"
                                    },
                                    "subject_id": {
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "contest_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Question updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "questions"
                ],
                "summary": "Remove the specified resource from storage.",
                "description": "remove a question on database",
                "operationId": "App\\Http\\Controllers\\QuestionController::destroy",
                "parameters": [
                    {
                        "name": "question",
                        "in": "path",
                        "description": "Question id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Question deleted"
                    }
                }
            }
        },
...
        "/subjects/{subject}": {
            "get": {
                "tags": [
                    "subjects"
                ],
                "summary": "Display the specified resource.",
                "description": "show subject",
                "operationId": "App\\Http\\Controllers\\SubjectController::show",
                "parameters": [
                    {
                        "name": "subject",
                        "in": "path",
                        "description": "Subject id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "only",
                        "in": "query",
                        "description": "filter results using field1;field2;field3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "with",
                        "in": "query",
                        "description": "get relations using relation1;relation2;relation3...",
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show subject"
                    }
                }
            },
            "put": {
                "tags": [
                    "subjects"
                ],
                "summary": "Update the specified resource in storage",
                "description": "update a subject on database",
                "operationId": "App\\Http\\Controllers\\SubjectController::update",
                "parameters": [
                    {
                        "name": "subject",
                        "in": "path",
                        "description": "Subject id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "subcategory_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Subject updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "subjects"
                ],
                "summary": "Remove the specified resource from storage.",
                "description": "remove a subject on database",
                "operationId": "App\\Http\\Controllers\\SubjectController::destroy",
                "parameters": [
                    {
                        "name": "subject",
                        "in": "path",
                        "description": "Subject id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subject deleted"
                    }
                }
            }
        }
    }
}